+++ feature/follower/src/main/java/com/kolown/follower/navigation/FollowerNavigation.kt
@@ -1,22 +1,22 @@
-package com.porring.camera.navigation
+package com.kolown.follower.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.porring.camera.CameraRoute
+import com.kolown.follower.FollowerRoute
 import com.porring.navigation.MainMenuRoute
 
-fun NavController.navigateCamera(navOptions: NavOptions) {
-    navigate(MainMenuRoute.Camera, navOptions)
+fun NavController.navigateFollower(navOptions: NavOptions) {
+    navigate(MainMenuRoute.Follower, navOptions)
 }
 
-fun NavGraphBuilder.cameraNavGraph(
+fun NavGraphBuilder.followerNavGraph(
     padding: PaddingValues,
 ) {
-    composable<MainMenuRoute.Camera> {
-        CameraRoute(
+    composable<MainMenuRoute.Follower> {
+        FollowerRoute(
             padding = padding
         )
     }
