+++ feature/home/src/main/java/com/kolown/home/navigation/HomeNavigation.kt
@@ -1,11 +1,11 @@
-package com.porring.home.navigation
+package com.kolown.home.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
-import com.porring.home.HomeRoute
+import com.kolown.home.HomeRoute
 import com.porring.navigation.MainMenuRoute
 
 fun NavController.navigateHome(navOptions: NavOptions) {
