+++ feature/search/src/main/java/com/kolown/search/navigation/SearchNavigation.kt
@@ -1,4 +1,4 @@
-package com.porring.search.navigation
+package com.kolown.search.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.navigation.NavController
@@ -6,7 +6,7 @@ import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import com.porring.navigation.MainMenuRoute
-import com.porring.search.SearchRoute
+import com.kolown.search.SearchRoute
 
 fun NavController.navigateSearch(navOptions: NavOptions) {
     navigate(MainMenuRoute.Search, navOptions)
